LayoutTests/js/arrowfunction-superproperty-expected.txt

Tests for ES6 arrow function, access to the super property in arrow function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (new B()).getValueParentFunction() is expectedValue
PASS (new C(false)).value is expectedValue
PASS E.getParentStaticValue() is expectedValue
PASS f.prop is expectedValue + "-" + expectedValue
PASS f.prop is expectedValue + "-" + "new-value"
PASS (new F()).getParentValue() is expectedValue
PASS successfullyParsed is true

TEST COMPLETE
